image_processing/requirements.txt
fastapi[standard]==0.115.12
httpx==0.28.1
psycopg2-binary==2.9.10
pillow==11.2.1
numpy==2.2.0
pydantic==2.11.3